Convert adsenseLoader util to TypeScript
